Took parameter sampler and mapping function out into different files to make it more modular and readable
